refactor: split promote gate cli parsing
